Replace usize with u8.

+++ src/event.rs
@@ -1,9 +1,9 @@
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum Event {
-    Pressed(usize),
-    Press(usize),
-    Release(usize),
-    Released(usize),
+    Pressed(u8),
+    Press(u8),
+    Release(u8),
+    Released(u8),
 }
 
 impl Default for Event {
